move cost_update_service to its own module; replay_stage is now decoupled with cost_model.
