+++ core/src/lib.rs
@@ -20,6 +20,7 @@ pub mod completed_data_sets_service;
 pub mod consensus;
 pub mod cost_model;
 pub mod cost_tracker;
+pub mod cost_update_service;
 pub mod execute_cost_table;
 pub mod fetch_stage;
 pub mod fork_choice;
